Separate cursor logic.
